requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.14.0
mysql-connector==2.2.9
mysql-connector-python==8.4.0
pytz==2024.1
sqlparse==0.5.0